+++ src/sha256.ts
@@ -224,14 +224,14 @@ export class Hash {
     }
 
     // Internal function for use in HMAC for optimization.
-    _saveState(out: Uint8Array) {
+    _saveState(out: Uint32Array) {
         for (let i = 0; i < this.state.length; i++) {
             out[i] = this.state[i];
         }
     }
 
     // Internal function for use in HMAC for optimization.
-    _restoreState(from: Uint8Array, bytesHashed: number) {
+    _restoreState(from: Uint32Array, bytesHashed: number) {
         for (let i = 0; i < this.state.length; i++) {
             this.state[i] = from[i];
         }
